storage_traits

Module indexeddb

Structs

DatabaseInfo

IndexedDBIndex

IndexedDBKeyRange

IndexedDBObjectStore

IndexedDBRecord

https://w3c.github.io/IndexedDB/#record-snapshot

TxnCompleteMsg

Enums

AsyncOperation

Operations that are not executed instantly, but rather added to a queue that is eventually run.

AsyncReadOnlyOperation

AsyncReadWriteOperation

AsyncSchemaOperation

BackendError

Any error from the backend, a super-set of DbError

ConnectionMsg

Messaging used in the context of connection lifecycle management.

CreateObjectResult

IndexedDBKeyType

https://www.w3.org/TR/IndexedDB-3/#key-type

IndexedDBThreadMsg

IndexedDBTxnMode

KeyPath

PutItemResult

SyncOperation

Type Aliases

BackendResult

DbError

DbResult

A DbResult wraps any part of a call that has to reach into the backend (in this case sqlite.rs) These errors could be anything, depending on the backend