storage_traits

Module indexeddb

Structs

DatabaseInfo

IndexedDBKeyRange

IndexedDBRecord

https://w3c.github.io/IndexedDB/#record-snapshot

Enums

AsyncOperation

Operations that are not executed instantly, but rather added to a queue that is eventually run.

AsyncReadOnlyOperation

AsyncReadWriteOperation

BackendError

Any error from the backend, a super-set of DbError

CreateObjectResult

IndexedDBKeyType

https://www.w3.org/TR/IndexedDB-2/#key-type

IndexedDBThreadMsg

IndexedDBTxnMode

KeyPath

OpenDatabaseResult

Result of https://w3c.github.io/IndexedDB/#open-a-database-connection Upgrade is an optional first message, followed by the ultimate result. TODO: refactor messaging into a coherent whole and with a persistent communication mechanism.

PutItemResult

SyncOperation

Type Aliases

BackendResult

DbError

DbResult

A DbResult wraps any part of a call that has to reach into the backend (in this case sqlite.rs) These errors could be anything, depending on the backend